Update patch set 1
Patch Set 1:

It should be noted that it is not a prerequisite to be a Nokia employee in order to be a committer, so it would be fine for Kristian to continue working on and being a committer if he wishes to, especially if he finds a new job where his Akraino experience is of use. But we can handle that if/when he indicates such a wish.

Patch-set: 1